Add files via upload
It is mainly a package for Java that you can use for replacing, printing, or removing vowels from a string. As we all know, replacing, printing, or removing vowels from a string is a long and hectic procedure. However, with the usage of this package, you can do it with a single line of code.